refactor(scrum-server): refactor user api
