+++ packages/scrum-server/bin/www
@@ -23,7 +23,7 @@ app.set('port', port);
 
 var server = http.createServer(app);
 
-var SocketService = require('../utils/socket-singleton');
+var SocketService = require('../src/services/socket/socket-singleton');
 app.set("socketService", new SocketService(server));
 
 /**
