+++ packages/scrum-server/src/api/components/user/controller.js
@@ -1,0 +1,2 @@
+// controller.ts
+// The controller class handles incoming requests, validates them and sends the response data back to the client. It uses the service class to interact with the database.
