+++ packages/scrum-server/src/api/components/user/index.js
@@ -1,0 +1,5 @@
+const User = require('./model');
+const routes = require('./routes');
+
+exports.User = User;
+exports.routes = routes;
