+++ packages/scrum-server/src/services/authStrategies/localStrategy.js
@@ -1,7 +1,7 @@
 /* eslint-disable */
 const LocalStrategy = require('passport-local').Strategy;
 const bcrypt = require('bcrypt');
-const User = require('../models/user');
+const User = require('../src/api/components/user/model');
 
 module.exports = function (passport) {
   passport.use(
